backend/requirements.txt
Django==2.2.8
djangorestframework==3.10.3
django-cors-headers==3.1.0
django-rest-swagger==2.2.0
pymongo==3.9.0
pandas